sim/requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
cloudpickle==2.0.0
cryptography==3.4.8
docker==5.0.2
faiss-cpu==1.7.1.post2
Flask==2.0.1
Flask-And-Redis==1.0.0
Flask-RESTful==0.3.9
gym==0.20.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-dataclass==8.5.3
mmh3==3.0.0
mypy-extensions==0.4.3
numpy==1.21.2
pandas==1.3.3
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
scipy==1.7.1
scp==0.14.1
six==1.16.0
style==1.1.6
tabulate==0.8.9
tqdm==4.62.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
zipp==3.5.0
gevent==22.10.2